requirements.txt
Flask==2.0.2
PyYAML==6.0
